# Conflicts:
#	src/DuckFunkCore.php

+++ database/migrations/2019_11_25_000000_update_permissions_table.php
@@ -1,25 +1,30 @@
 <?php
 
+use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class UpdatePermissionsTable extends \Illuminate\Database\Migrations\Migration
+class UpdatePermissionsTable extends Migration
 {
     public function up()
     {
-        Schema::table('permissions', function (Blueprint $table) {
-            $table->boolean('duck_funk_housekeeping_read');
-            $table->boolean('duck_funk_housekeeping_write');
-            $table->boolean('duck_funk_housekeeping_approve');
-        });
+        if (Schema::hasTable('permissions')) {
+            Schema::table('permissions', function (Blueprint $table) {
+                $table->boolean('duck_funk_housekeeping_read');
+                $table->boolean('duck_funk_housekeeping_write');
+                $table->boolean('duck_funk_housekeeping_approve');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('permissions', function (Blueprint $table) {
-            $table->dropColumn('duck_funk_housekeeping_read');
-            $table->dropColumn('duck_funk_housekeeping_write');
-            $table->dropColumn('duck_funk_housekeeping_approve');
-        });
+        if (Schema::hasTable('permissions')) {
+            Schema::table('permissions', function (Blueprint $table) {
+                $table->dropColumn('duck_funk_housekeeping_read');
+                $table->dropColumn('duck_funk_housekeeping_write');
+                $table->dropColumn('duck_funk_housekeeping_approve');
+            });
+        }
     }
 }
